Conflicts:
	contrib/cmake/CMakeLists.txt
	src/MEGASync/MegaApplication.cpp
	src/MEGASync/MegaApplication.h
	src/MEGASync/gui/InfoDialog.h
	src/MEGASync/gui/MegaItem.cpp
	src/MEGASync/gui/MegaItemModel.cpp
	src/MEGASync/gui/MegaItemTreeView.cpp
	src/MEGASync/gui/NodeSelector.cpp
	src/MEGASync/gui/win/InfoDialog.ui
	src/MEGASync/model/Model.h
	src/MEGASync/platform/linux/LinuxPlatform.cpp

+++ src/MEGASync/google_breakpad/client/linux/log/log.cc
@@ -41,7 +41,7 @@ int write(const char* buf, size_t nbytes) {
 #if defined(__ANDROID__)
   return __android_log_write(ANDROID_LOG_WARN, "google-breakpad", buf);
 #else
-  return sys_write(2, buf, nbytes);
+  return static_cast<int>(sys_write(2, buf, nbytes));
 #endif
 }
